Refactor AboutActivity
