+++ app/src/main/java/com/chrisvasqm/cuadramo/about/AboutActivity.kt
@@ -1,15 +1,16 @@
 package com.chrisvasqm.cuadramo.about
 
+import android.app.Activity
+import android.content.Intent
+import android.net.Uri
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.chrisvasqm.cuadramo.databinding.ActivityAboutBinding
 import kotlinx.android.synthetic.main.toolbar.*
 
-class AboutActivity : AppCompatActivity(), AboutContract.View {
+class AboutActivity : AppCompatActivity() {
 
     private lateinit var binding: ActivityAboutBinding
-
-    private lateinit var router: AboutContract.Router
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -23,20 +24,27 @@ class AboutActivity : AppCompatActivity(), AboutContract.View {
         binding.linearGitHubLink.setOnClickListener { goToGitHub() }
         binding.linearChristianLink.setOnClickListener { goToChristianLink() }
         binding.linearCarlaLink.setOnClickListener { goToCristalLink() }
-
-        router = AboutRouter(this)
     }
 
-    override fun goToGitHub() {
-        router.goToGitHub()
+    private fun goToGitHub() {
+        val url = "https://github.com/chrisvasqm/cuadramo"
+        goToLink(this, url)
     }
 
-    override fun goToChristianLink() {
-        router.goToChristianLink()
+    private fun goToChristianLink() {
+        val url = "https://github.com/chrisvasqm"
+        goToLink(this, url)
     }
 
-    override fun goToCristalLink() {
-        router.goToCristalLink()
+    private fun goToCristalLink() {
+        val url = "https://www.behance.net/hipandac"
+        goToLink(this, url)
+    }
+
+    private fun goToLink(activity: Activity, url: String) {
+        Intent(Intent.ACTION_VIEW)
+            .apply { data = Uri.parse(url) }
+            .also { activity.startActivity(it) }
     }
 
     private fun displayVersionNumber() {
