--- app/src/main/java/com/chrisvasqm/cuadramo/about/AboutContract.kt
@@ -1,25 +1,0 @@
-package com.chrisvasqm.cuadramo.about
-
-interface AboutContract {
-
-    interface View {
-
-        fun goToGitHub()
-
-        fun goToChristianLink()
-
-        fun goToCristalLink()
-
-    }
-
-    interface Router {
-
-        fun goToGitHub()
-
-        fun goToChristianLink()
-
-        fun goToCristalLink()
-
-    }
-
-}
